Support widening primitive conversion for method invocations
This commit introduces support in ReflectionUtils.isAssignableTo() for
determining if Java supports a "widening primitive conversion" from the
source object to the target type.

In addition, the DefaultArgumentConverter for implicit argument
conversion in parameterized tests has also been modified to use
ReflectionUtils.isAssignableTo() instead of performing a simple
"instance of" check against the target type. Consequently, primitive
arguments (or their corresponding wrappers) may now be supplied to a
@ParameterizedTest with automatic widening conversion.

Issue: #1092